+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutLineItem.java
@@ -29,22 +29,44 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
     public static final String KIND_CREDIT = "credit";
     public static final String KIND_DEBIT = "debit";
 
+    /**
+     * The upc type of PayPal line item.
+     */
+    @Retention(RetentionPolicy.SOURCE)
+    @StringDef({PayPalNativeCheckoutLineItem.UPC_TYPE_A, PayPalNativeCheckoutLineItem.UPC_TYPE_B, PayPalNativeCheckoutLineItem.UPC_TYPE_C, PayPalNativeCheckoutLineItem.UPC_TYPE_D, PayPalNativeCheckoutLineItem.UPC_TYPE_E, PayPalNativeCheckoutLineItem.UPC_TYPE_2, PayPalNativeCheckoutLineItem.UPC_TYPE_5})
+    @interface PayPalLineItemUpcType {
+    }
+
+    public static final String UPC_TYPE_A = "UPC-A";
+    public static final String UPC_TYPE_B = "UPC-B";
+    public static final String UPC_TYPE_C = "UPC-C";
+    public static final String UPC_TYPE_D = "UPC-D";
+    public static final String UPC_TYPE_E = "UPC-E";
+    public static final String UPC_TYPE_2 = "UPC-2";
+    public static final String UPC_TYPE_5 = "UPC-5";
+
     private static final String DESCRIPTION_KEY = "description";
+    private static final String IMAGE_URL_KEY = "image_url";
     private static final String KIND_KEY = "kind";
     private static final String NAME_KEY = "name";
     private static final String PRODUCT_CODE_KEY = "product_code";
     private static final String QUANTITY_KEY = "quantity";
     private static final String UNIT_AMOUNT_KEY = "unit_amount";
     private static final String UNIT_TAX_AMOUNT_KEY = "unit_tax_amount";
+    private static final String UPC_CODE_KEY = "upc_code";
+    private static final String UPC_TYPE_KEY = "upc_type";
     private static final String URL_KEY = "url";
 
     private String description;
+    private String imageUrl;
     private String kind;
     private String name;
     private String productCode;
     private String quantity;
     private String unitAmount;
     private String unitTaxAmount;
+    private String upcCode;
+    private String upcType;
     private String url;
 
     /**
@@ -72,6 +94,15 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
      */
     public void setDescription(@NonNull String description) {
         this.description = description;
+    }
+
+    /**
+     * The image URL to product information.
+     *
+     * @param imageUrl The image URL with additional information.
+     */
+    public void setImageUrl(@Nullable String imageUrl) {
+        this.imageUrl = imageUrl;
     }
 
     /**
@@ -129,6 +160,24 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
     }
 
     /**
+     * UPC code of the item.
+     *
+     * @param upcCode The UPC code.
+     */
+    public void setUpcCode(@Nullable String upcCode) {
+        this.upcCode = upcCode;
+    }
+
+    /**
+     * UPC type of the item.
+     *
+     * @param upcType The UPC type.
+     */
+    public void setUpcType(@Nullable @PayPalLineItemUpcType String upcType) {
+        this.upcType = upcType;
+    }
+
+    /**
      * The URL to product information.
      *
      * @param url The URL with additional information.
@@ -140,6 +189,11 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
     @Nullable
     public String getDescription() {
         return description;
+    }
+
+    @Nullable
+    public String getImageUrl() {
+        return imageUrl;
     }
 
     @PayPalLineItemKind
@@ -174,6 +228,17 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
     }
 
     @Nullable
+    public String getUpcCode() {
+        return upcCode;
+    }
+
+    @PayPalLineItemUpcType
+    @Nullable
+    public String getUpcType() {
+        return upcType;
+    }
+
+    @Nullable
     public String getUrl() {
         return url;
     }
@@ -182,12 +247,15 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
         try {
             return new JSONObject()
                     .putOpt(DESCRIPTION_KEY, description)
+                    .putOpt(IMAGE_URL_KEY, imageUrl)
                     .putOpt(KIND_KEY, kind)
                     .putOpt(NAME_KEY, name)
                     .putOpt(PRODUCT_CODE_KEY, productCode)
                     .putOpt(QUANTITY_KEY, quantity)
                     .putOpt(UNIT_AMOUNT_KEY, unitAmount)
                     .putOpt(UNIT_TAX_AMOUNT_KEY, unitTaxAmount)
+                    .putOpt(UPC_CODE_KEY, upcCode)
+                    .putOpt(UPC_TYPE_KEY, upcType)
                     .putOpt(URL_KEY, url);
         } catch (JSONException ignored) {
         }
@@ -197,12 +265,15 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
 
     PayPalNativeCheckoutLineItem(Parcel in) {
         description = in.readString();
+        imageUrl = in.readString();
         kind = in.readString();
         name = in.readString();
         productCode = in.readString();
         quantity = in.readString();
         unitAmount = in.readString();
         unitTaxAmount = in.readString();
+        upcCode = in.readString();
+        upcType = in.readString();
         url = in.readString();
     }
 
@@ -226,12 +297,15 @@ public class PayPalNativeCheckoutLineItem implements Parcelable {
     @Override
     public void writeToParcel(Parcel parcel, int i) {
         parcel.writeString(description);
+        parcel.writeString(imageUrl);
         parcel.writeString(kind);
         parcel.writeString(name);
         parcel.writeString(productCode);
         parcel.writeString(quantity);
         parcel.writeString(unitAmount);
         parcel.writeString(unitTaxAmount);
+        parcel.writeString(upcCode);
+        parcel.writeString(upcType);
         parcel.writeString(url);
     }